fix bugs and add some features

+++ src/main/java/vn/edu/iuh/fit/se/thubonggiareapis/service/impl/UserServiceImpl.java
@@ -1,18 +1,26 @@
 package vn.edu.iuh.fit.se.thubonggiareapis.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import vn.edu.iuh.fit.se.thubonggiareapis.converter.UserConverter;
 import vn.edu.iuh.fit.se.thubonggiareapis.dto.UserDTO;
 import vn.edu.iuh.fit.se.thubonggiareapis.entity.User;
 import vn.edu.iuh.fit.se.thubonggiareapis.repository.UserRepository;
+import vn.edu.iuh.fit.se.thubonggiareapis.service.IPasswordService;
 import vn.edu.iuh.fit.se.thubonggiareapis.service.IUserService;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class UserServiceImpl implements IUserService {
+
+    @Autowired
+    private IPasswordService passwordService;
 
     @Autowired
     private UserRepository userRepository;
@@ -23,13 +31,14 @@ public class UserServiceImpl implements IUserService {
     @Override
     public UserDTO addUser(UserDTO userDTO) {
         User user = userConverter.toEntity(userDTO);
+        user.setPassword(passwordService.passwordEncoder().encode(user.getPassword()));
         user = userRepository.save(user);
         return userConverter.toDto(user);
     }
 
     @Override
     public List<UserDTO> getUsers() {
-        List<UserDTO> userDTOs = new ArrayList<UserDTO>();
+        List<UserDTO> userDTOs = new ArrayList<>();
         List<User> users = userRepository.findAll();
         for (User user : users) {
             UserDTO userDTO = userConverter.toDto(user);
@@ -40,23 +49,26 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public UserDTO getUser(long id) {
-        User user = userRepository.getById(id);
-        UserDTO userDTO = userConverter.toDto(user);
-        return userDTO;
+        User user = userRepository.findById(id).orElse(null);
+        if (Objects.isNull(user)) {
+            return null;
+        }
+        return userConverter.toDto(user);
     }
 
     @Override
     public void updateUser(UserDTO userDTO) {
-        User user = userConverter.toEntity(userDTO);
-        userRepository.save(user);
-
+        if (userRepository.existsById(userDTO.getId())) {
+            User user = userConverter.toEntity(userDTO);
+            user.setPassword(passwordService.passwordEncoder().encode(user.getPassword()));
+            userRepository.save(user);
+        }
     }
 
     @Override
     public void deleteUser(long id) {
-        userRepository.deleteById(id);
-
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        }
     }
-
-
 }
